+++ test/Script.cpp
@@ -6,10 +6,14 @@
  */
 
 #include <iostream>
+#include <cstring>
+#include <cstdlib>
+#include <zlib.h>
 
 #include "global.h"
 #include "utils/strings.h"
 #include "Script.h"
+#include "ScriptTest.h"
 
 #define printStrings(cnt, acc, name) { \
     std::cout << name << " (" << cnt << ")" << std::endl; \
@@ -44,10 +48,7 @@
 }
 
 namespace {
-    int test(const char *f, bool strings) {
-        Script s;
-        assertEqual(s.load(f), 0);
-
+    int printDataScript(Script &s, bool strings) {
         if (strings) {
             printStrings(s.levelCount(), s.getLevelName, "Level Names");
             printStrings(s.levelCount(), s.getLevelFilename, "Level Filenames");
@@ -92,46 +93,142 @@ namespace {
                 else
                     std::cout << "Script for \"" << s.getLevelName(i - 1) << "\" (" << i - 1 << "):" << std::endl;
                 int error = s.runScript(i);
-                if (error != 0)
+                if (error != 0) {
                     std::cout << "Returned " << error << "..." << std::endl;
+                    return error;
+                }
                 std::cout << std::endl;
             }
         }
 
         return 0;
     }
+
+    int test(const char *file, unsigned int n) {
+        Script s;
+
+        std::cout << "Testing " << testDescription[n] << std::endl;
+
+        if (s.load(file) != 0) {
+            std::cout << "Could not open file " << file << std::endl;
+            return 1;
+        }
+
+        if (s.gameStringCount() != testExpectedGameStringCount[n]) {
+            std::cout << "Game String Count " << s.gameStringCount() << " != " << testExpectedGameStringCount[n] << std::endl;
+            return 2;
+        }
+
+        if (s.pcStringCount() != testExpectedPlatformStringCount[n]) {
+            std::cout << "Platform String Count " << s.pcStringCount() << " != " << testExpectedPlatformStringCount[n] << std::endl;
+            return 3;
+        }
+
+        std::cout << "Success!" << std::endl << std::endl;
+        return 0;
+    }
+
+    void readPayloadChunk(const unsigned char *data, unsigned int size, const char *file) {
+        static const unsigned int bufferSize = 16384; // 16K should be enough for everybody :)
+        unsigned char buffer[bufferSize];
+
+        z_stream stream;
+        stream.zalloc = Z_NULL;
+        stream.zfree =  Z_NULL;
+        stream.opaque = Z_NULL;
+        assertEqual(inflateInit2(&stream, 16), Z_OK); // 16 -> gzip header
+
+        // Inflate data in one go
+        stream.avail_in = size;
+        stream.next_in = const_cast<unsigned char *>(data);
+        stream.avail_out = bufferSize;
+        stream.next_out = buffer;
+        assertEqual(inflate(&stream, Z_FINISH), Z_STREAM_END);
+        inflateEnd(&stream);
+
+        // Write buffer to file
+        std::ofstream s(file, std::ios_base::out | std::ios_base::binary);
+        s.write(reinterpret_cast<const char *>(buffer), bufferSize - stream.avail_out);
+    }
+
+    int runForPayload(unsigned int n, bool print, bool printData) {
+        assert(n < testPayloadCount);
+        // Get temp file name
+        char tmpFile[] = "/tmp/openraider_unit_test_0";
+        FILE *f;
+        while ((f = fopen(tmpFile, "r")) != NULL) {
+            fclose(f);
+            tmpFile[25]++;
+        }
+
+        std::cout << "Temporary test file: " << tmpFile << std::endl;
+
+        readPayloadChunk(testPayloads[n], testSizes[n], tmpFile);
+
+        int error = 0;
+        if (print) {
+            Script s;
+            error = s.load(tmpFile);
+            if (error == 0)
+                error = printDataScript(s, printData);
+        } else {
+            error = test(tmpFile, n);
+        }
+
+        remove(tmpFile);
+        return error;
+    }
 }
 
 int main(int argc, char *argv[]) {
-    std::cout << "Tomb Raider 2:" << std::endl;
-    char *f = fullPath("~/.OpenRaider/paks/tr2/TOMBPC.DAT", 0);
-    int error = test(f, !((argc > 1) && (argv[1][0] == 's')));
-    delete [] f;
+    bool printHelp = false;
+    bool print = false;
+    bool printData = true;
+    int whichFile = -1;
 
-    if (error != 0)
-        return error;
+    if (argc == 3) {
+        if ((strcmp(argv[1], "--printData") == 0)
+                || (strcmp(argv[1], "--printScript") == 0)) {
+            print = true;
+            if (strcmp(argv[1], "--printScript") == 0) {
+                printData = false;
+            }
+            assert(testPayloadCount < 10);
+            if ((argv[2][0] >= '0') && ((unsigned int)argv[2][0] <= (testPayloadCount + '0'))) {
+                whichFile = argv[2][0] - '0';
+            }
+        } else {
+            printHelp = true;
+        }
+    } else if (argc != 1) {
+        printHelp = true;
+    }
 
-    std::cout << std::endl << "Tomb Raider 3:" << std::endl;
-    f = fullPath("~/.OpenRaider/paks/tr3/TOMBPC.DAT", 0);
-    error = test(f, !((argc > 1) && (argv[1][0] == 's')));
-    delete [] f;
+    if (printHelp) {
+        std::cout << "Usage:" << std::endl;
+        std::cout << "\t" << argv[0] << " [--printData | --printScript] [N | /path]" << std::endl;
+        return 1;
+    }
 
-    if (error != 0)
-        return error;
-
-    std::cout << std::endl << "Tomb Raider 2 PSX:" << std::endl;
-    f = fullPath("~/.OpenRaider/paks/tr2_psx/TOMBPSX.DAT", 0);
-    error = test(f, !((argc > 1) && (argv[1][0] == 's')));
-    delete [] f;
-
-    if (error != 0)
-        return error;
-
-    std::cout << std::endl << "Tomb Raider 3 PSX:" << std::endl;
-    f = fullPath("~/.OpenRaider/paks/tr3_psx/TOMBPSX.DAT", 0);
-    error = test(f, !((argc > 1) && (argv[1][0] == 's')));
-    delete [] f;
-
-    return error;
+    if (print) {
+        // Print single script
+        if (whichFile == -1) {
+            // From given path
+            Script s;
+            assertEqual(s.load(argv[2]), 0);
+            return printDataScript(s, printData);
+        } else {
+            // From payload
+            return runForPayload((unsigned int)whichFile, true, printData);
+        }
+    } else {
+        // Run test on all scripts in payload
+        for (unsigned int i = 0; i < testPayloadCount; i++) {
+            int error = runForPayload(i, false, false);
+            if (error != 0)
+                return error;
+        }
+        return 0;
+    }
 }